+++ lang/runtime/error_dummy.h
@@ -1,10 +1,12 @@
 // Minimal dummy error constructor for SSA error-path testing.
 #pragma once
 
+#include <stdint.h>
+
 struct DriftError {
-    long long code; // matches Drift Int (word-sized)
+    int64_t code; // matches Drift Int (word-sized)
 };
 
 // Returns a non-null Error* for testing error-edge lowering.
-struct DriftError* drift_error_new_dummy(int code);
-long long drift_error_get_code(struct DriftError* err);
+struct DriftError* drift_error_new_dummy(int64_t code);
+int64_t drift_error_get_code(struct DriftError* err);
